Перенос views локаций в папку приложения
